feat: genesis of the v2 user-first SDK :woohooo:
the big changes here are updating the func signatures to remove confusion
with error path, APIErrors, and logging. Additionally, we remove much of
the start up issues we've had previously, with the logger/config setup, and
move that to the new fdk.Run func, which utilizes generics to resolve all
the issues we've been seeing pop from authors around the "fn isn't working"
notices we've had. All in all, we're trying to reduce the boilerplate, remove
bullets from the boostrappping footgun, and provide a more ergonomic
experience for the func authors.