chore: drop falcon client dependency and replace with helpers and doc help
we've effectively had the dependency relationship inverted here. If anything
we'll want to include the intracies of the fdk, which are minimal, in the
github.com/CrowdStrike/gofalcon module, not the other way around. This
allows all faas authors to enjoy the lightweight lib that is fdk without
carrying the compilation/security/breadth concerns.