fix: drop Fn Version from Fn output
this can create some issues with startup sequencing. Ideallly, the user
is not concerning themselves with the fn version in any way. We can mostly
eliminate that. At the moment, we have no ask for such a thing. Lets rip
it out for now then add it back when there is an ask.

Its likely better off as a request input rather than an env var. There's
no guarantee a fn version creates a separate lambda deploy. By adding it
to the request, we eliminate that confusion altogether.