chore: add secondary means of testing
I quite like this latter setup. We get some simple fns to work
with, and these can be combined in a myriad of ways. One big
hole we have at the moment is testing with the runner's behavior
in the mix. There are things the runner does that needs to be
accounted for. Most tests do not account for it atm.

With these test fns, we could create a helper that does all
the setup/runner stuff and allows the user to write dead simple
tests.